alterando o valor default do formulario de cadastro de produto

+++ src/components/produtos/deletarProduto/DeletarProduto.tsx
@@ -59,14 +59,20 @@ function DeletarProduto() {
 
   function sim() {
 
+    if (produto?.usuario?.id == parseInt(idUser)) {
 
-    history.push('/produtos')
-    deleteId(`/produto/${id}`, {
-      headers: {
-        'Authorization': token
-      }
-    });
-    alert('Produto deletado com sucesso');
+      history.push('/produtos')
+      deleteId(`/produto/${id}`, {
+        headers: {
+          'Authorization': token
+        }
+      });
+      alert('Produto deletado com sucesso');
+    }
+    else {
+      alert("Você não pode deletar este produto")
+      history.push('/produtos');
+    }
   }
 
 
@@ -82,7 +88,7 @@ function DeletarProduto() {
           <CardContent>
             <Box justifyContent="center">
               <Typography color="textSecondary" gutterBottom>
-                {usuarios.map(usuario => usuario.id == parseInt(idUser) ? usuario.nome : "")}, tem certeza que deseja deletar o {produto?.servico==true?"produto":"serviço"}:
+                {usuarios.map(usuario => usuario.id == parseInt(idUser) ? usuario.nome : "")}, tem certeza que deseja deletar o {produto?.servico == true ? "produto" : "serviço"}:
               </Typography>
               <Typography color="textSecondary">
                 {produto?.nome}
